Tests

+++ src/modules/Category/deleteCategory/DeleteCategoryService.ts
@@ -4,6 +4,10 @@ export class DeleteCategoryService {
   constructor(private categoryRepository: ICategoryRepository) {}
 
   async execute(id: string): Promise<Error | void> {
+    const itemExistsCategory = await this.categoryRepository.existsItem(id);
+    if (itemExistsCategory) {
+      return new Error("This category has linked items and cannot be deleted");
+    }
     const result = await this.categoryRepository.delete(id);
 
     if (result == false) {
